Minor refactor and cleanup
This will simplify some planned work
