+++ martin/src/lib.rs
@@ -26,8 +26,10 @@ mod test_utils;
 #[cfg(test)]
 pub use crate::args::Env;
 pub use crate::config::{read_config, Config};
-pub use crate::source::{IdResolver, Source, Sources, Xyz};
-pub use crate::utils::{decode_brotli, decode_gzip, BoolOrObject, Error, OneOrMany, Result};
+pub use crate::source::{Source, Sources, Xyz};
+pub use crate::utils::{
+    decode_brotli, decode_gzip, BoolOrObject, Error, IdResolver, OneOrMany, Result,
+};
 
 // Ensure README.md contains valid code
 #[cfg(doctest)]
