+++ martin/src/srv/mod.rs
@@ -2,4 +2,4 @@ mod config;
 mod server;
 
 pub use config::{SrvConfig, KEEP_ALIVE_DEFAULT, LISTEN_ADDRESSES_DEFAULT};
-pub use server::{new_server, router, AppState, IndexEntry, RESERVED_KEYWORDS};
+pub use server::{new_server, router, RESERVED_KEYWORDS};
