+++ martin/src/utils/mod.rs
@@ -1,7 +1,9 @@
 mod error;
+mod id_resolver;
 mod one_or_many;
 mod utilities;
 
 pub use error::*;
+pub use id_resolver::IdResolver;
 pub use one_or_many::OneOrMany;
 pub use utilities::*;
